refactor: move types to a single file
