--- packages/pme/src/meta/base.ts
@@ -1,8 +1,0 @@
-export enum MetaKind {
-  extensionLink = 'extensionLink',
-  extensionComponent = 'extensionComponent',
-}
-
-export interface MetaBase {
-  kind: MetaKind;
-}
